[MIG] stock_usability to v14
