+++ stock_usability/__init__.py
@@ -1,7 +1,2 @@
-# -*- coding: utf-8 -*-
-
-from . import stock
-from . import procurement
-from . import product
-from . import partner
+from . import models
 from . import wizard
