+++ stock_usability/__manifest__.py
@@ -1,11 +1,11 @@
-# Copyright 2014-2019 Akretion (http://www.akretion.com)
+# Copyright 2014-2020 Akretion (http://www.akretion.com)
 # @author Alexis de Lattre <alexis.delattre@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 
 {
     'name': 'Stock Usability',
-    'version': '12.0.1.0.0',
+    'version': '14.0.1.0.0',
     'category': 'Inventory, Logistic, Storage',
     'license': 'AGPL-3',
     'summary': 'Several usability enhancements in Warehouse management',
@@ -19,7 +19,6 @@ The usability enhancements include:
 * add a group by Partner in the picking search view (particularly usefull for receptions)
 * add graph view for pickings
 * remove ability to translate stock.location, stock.location.route and stock.picking.type
-* unactive view "stock.view_production_lot_form_simple" because it doesn't include smart buttons Locate and Traceability Report ; then we only use the main view "stock.view_production_lot_form" like in previous odoo versions
 
 This module has been written by Alexis de Lattre from Akretion <alexis.delattre@akretion.com>.
     """,
@@ -27,10 +26,15 @@ This module has been written by Alexis de Lattre from Akretion <alexis.delattre@
     'website': 'http://www.akretion.com',
     'depends': ['stock'],
     'data': [
-        'stock_view.xml',
-        'procurement_view.xml',
-        'wizard/stock_quantity_history_view.xml',
+        'views/stock_quant.xml',
+        'views/stock_inventory.xml',
+        'views/stock_location.xml',
+        'views/stock_move.xml',
+        'views/stock_picking.xml',
+        'views/stock_warehouse.xml',
+        'views/procurement_group.xml',
+        'views/procurement_scheduler_log.xml',
         'security/ir.model.access.csv',
         ],
-    'installable': False,
+    'installable': True,
 }
