+++ stock_usability/models/__init__.py
@@ -1,0 +1,9 @@
+from . import stock_move
+from . import stock_picking
+from . import stock_location_route
+from . import stock_warehouse_orderpoint
+from . import stock_quant
+from . import procurement_group
+from . import procurement_scheduler_log
+from . import product_template
+from . import res_partner
