+++ stock_usability/models/procurement_group.py
@@ -1,8 +1,8 @@
-# Copyright 2015-2019 Akretion (http://www.akretion.com)
+# Copyright 2015-2020 Akretion (http://www.akretion.com)
 # @author Alexis de Lattre <alexis.delattre@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo import models, fields, api
+from odoo import api, fields, models
 from datetime import datetime
 import logging
 
@@ -12,29 +12,18 @@ logger = logging.getLogger(__name__)
 class ProcurementGroup(models.Model):
     _inherit = 'procurement.group'
 
-    @api.model
-    def _procure_orderpoint_confirm(
-            self, use_new_cursor=False, company_id=False):
-        logger.info(
-            'procurement scheduler: START to create moves from '
-            'orderpoints')
-        res = super(ProcurementGroup, self)._procure_orderpoint_confirm(
-            use_new_cursor=use_new_cursor, company_id=company_id)
-        logger.info(
-            'procurement scheduler: END creation of moves from '
-            'orderpoints')
-        return res
+    picking_ids = fields.One2many(
+        'stock.picking', 'group_id', string='Pickings', readonly=True)
 
     @api.model
-    def run_scheduler(
-            self, use_new_cursor=False, company_id=False):
+    def run_scheduler(self, use_new_cursor=False, company_id=False):
         '''Inherit to add info logs'''
         logger.info(
             'START procurement scheduler '
             '(company ID=%d, uid=%d, use_new_cursor=%s)',
             company_id, self._uid, use_new_cursor)
         start_datetime = datetime.now()
-        res = super(ProcurementGroup, self).run_scheduler(
+        res = super().run_scheduler(
             use_new_cursor=use_new_cursor, company_id=company_id)
         logger.info(
             'END procurement scheduler '
@@ -55,13 +44,3 @@ class ProcurementGroup(models.Model):
             logger.warning(
                 'Could not create procurement.scheduler.log (error: %s)', e)
         return res
-
-
-class ProcurementSchedulerLog(models.Model):
-    _name = 'procurement.scheduler.log'
-    _description = 'Logs of the Procurement Scheduler'
-    _order = 'create_date desc'
-
-    company_id = fields.Many2one(
-        'res.company', string='Company', readonly=True)
-    start_datetime = fields.Datetime(string='Start Date', readonly=True)
