+++ stock_usability/models/product_template.py
@@ -1,0 +1,12 @@
+# Copyright 2016-2020 Akretion France
+# @author: Alexis de Lattre <alexis.delattre@akretion.com>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class ProductTemplate(models.Model):
+    _inherit = 'product.template'
+
+    tracking = fields.Selection(tracking=True)
+    sale_delay = fields.Float(tracking=True)
