+++ stock_usability/models/res_partner.py
@@ -1,11 +1,11 @@
-# Copyright 2017-2019 Akretion France (https://akretion.com/)
+# Copyright 2017-2020 Akretion France (https://akretion.com/)
 # @author Alexis de Lattre <alexis.delattre@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo import models, fields
+from odoo import fields, models
 
 
 class ResPartner(models.Model):
     _inherit = 'res.partner'
 
-    picking_warn = fields.Selection(track_visibility='onchange')
+    picking_warn = fields.Selection(tracking=True)
