+++ stock_usability/models/stock_location_route.py
@@ -1,0 +1,11 @@
+# Copyright 2014-2020 Akretion (http://www.akretion.com)
+# @author Alexis de Lattre <alexis.delattre@akretion.com>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class StockLocationRoute(models.Model):
+    _inherit = 'stock.location.route'
+
+    name = fields.Char(translate=False)
