--- stock_usability/product.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-# © 2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from odoo import models, fields
-
-
-class ProductTemplate(models.Model):
-    _inherit = 'product.template'
-
-    tracking = fields.Selection(track_visibility='onchange')
-    sale_delay = fields.Float(track_visibility='onchange')
