+++ stock_usability/views/procurement_scheduler_log.xml
@@ -1,22 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  Copyright 2015-2019 Akretion (http://www.akretion.com/)
+  Copyright 2015-2020 Akretion (http://www.akretion.com/)
   @author Alexis de Lattre <alexis.delattre@akretion.com>
   License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 
 <odoo>
-
-<record id="procurement_group_form_view" model="ir.ui.view">
-    <field name="name">stock_usability.procurement.group.form</field>
-    <field name="model">procurement.group</field>
-    <field name="inherit_id" ref="stock.procurement_group_form_view"/>
-    <field name="arch" type="xml">
-        <field name="move_type" position="after">
-            <field name="partner_id" readonly="True"/>
-        </field>
-    </field>
-</record>
 
 <record id="procurement_scheduler_log_tree" model="ir.ui.view">
     <field name="name">procurement_scheduler_log_tree</field>
@@ -39,6 +28,6 @@
 
 <menuitem id="procurement_scheduler_log_menu"
     action="procurement_scheduler_log_action"
-    parent="stock.menu_stock_inventory_control" sequence="50"/>
+    parent="stock.menu_stock_warehouse_mgmt" sequence="140"/>
 
 </odoo>
