+++ stock_usability/views/stock_inventory.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  Copyright 2014-2020 Akretion (http://www.akretion.com/)
+  @author Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+
+<odoo>
+
+<record id="stock_inventory_line_tree" model="ir.ui.view">
+    <field name="name">usability.stock.inventory.line.tree</field>
+    <field name="model">stock.inventory.line</field>
+    <field name="inherit_id" ref="stock.stock_inventory_line_tree"/>
+    <field name="arch" type="xml">
+        <tree position="attributes">
+            <!-- native :
+            decoration-danger="product_qty != theoretical_qty"
+            decoration-muted="product_qty == theoretical_qty"
+            decoration-bf="is_editable"
+            -->
+            <attribute name="decoration-info">product_qty &gt; theoretical_qty</attribute>
+            <attribute name="decoration-danger">product_qty &lt; theoretical_qty</attribute>
+        </tree>
+    </field>
+</record>
+
+<record id="view_inventory_tree" model="ir.ui.view">
+    <field name="name">usability.stock.inventory.tree</field>
+    <field name="model">stock.inventory</field>
+    <field name="inherit_id" ref="stock.view_inventory_tree"/>
+    <field name="arch" type="xml">
+        <tree position="attributes">
+            <attribute name="decoration-info">state == 'draft'</attribute>
+            <attribute name="decoration-warning">state == 'confirm'</attribute>
+        </tree>
+    </field>
+</record>
+
+</odoo>
