--- stock_usability/wizard/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import stock_quantity_history
