--- stock_usability/wizard/stock_quantity_history_view.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  Copyright 2019 Akretion France (http://www.akretion.com/)
-  @author: Alexis de Lattre <alexis.delattre@akretion.com>
-  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
--->
-
-<odoo>
-
-
-<record id="view_stock_quantity_history" model="ir.ui.view">
-    <field name="name">stock_usability.stock.quantity.history.form</field>
-    <field name="model">stock.quantity.history</field>
-    <field name="inherit_id" ref="stock.view_stock_quantity_history"/>
-    <field name="arch" type="xml">
-        <field name="date" position="after">
-            <field name="location_id"/>
-        </field>
-    </field>
-</record>
-
-
-</odoo>
